Moved testing types into their own integration and unit test files. Updated old generic and td types testing and incorporated into new unit/test_types.py.
